# feat: faster loading if model already cached