This PR adds much faster model loading by directly loading from cache if a model is cached, if `force_download` is set to true during loading

Timings when loading `potion-base-8m`

* Before: 1.27 s ± 72.5 ms per loop
* Now: 37.3 ms ± 335 µs per loop

This speed is only enabled when setting `force_download` to `False`. It is true by default for backwards compatibility and download tracking

Mention for #275